Added velocity and variable radius
Size of bodies now depends on length of press and initial velocity
is determined by length (straight line) of gesture on screen.
