+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.planetgame">
+    package="com.riley.planetgame">
 
     <application
         android:allowBackup="true"
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.PlanetGame">
         <activity
-            android:name=".SimActivity"
+            android:name="com.riley.planetgame.SimActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.riley.planetgame.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
