+++ app/src/main/java/com/riley/planetgame/MainActivity.java
@@ -1,0 +1,31 @@
+package com.riley.planetgame;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import com.riley.planetgame.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    Button start_button;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        start_button = (Button)findViewById(R.id.start_button);
+        start_button.setOnClickListener(new View.OnClickListener(){
+           @Override
+           public void onClick(View v){
+               Intent intent = new Intent(MainActivity.this,
+                       SimActivity.class);
+               startActivity(intent);
+           }
+        });
+    }
+}
