+++ app/src/main/res/layout/activity_sim.xml
@@ -4,9 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SimActivity">
+    tools:context="com.riley.planetgame.SimActivity">
 
-    <com.example.planetgame.SimView
+    <com.riley.planetgame.SimView
         android:id="@+id/sim_view"
         android:layout_width="0dp"
         android:layout_height="0dp"
